requirements.txt
alembic==1.14.1
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.8
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lml==0.1.0
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pyexcel==0.7.1
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
reportlab==4.2.5
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
texttable==1.7.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1